[fix] : Restaurant ViewModel 분리
